datastar clouddnsservice xdata baseanti mongodb  (#23)
* publish jdcloud-sdk-go version 1.0.2

* publish jdcloud-sdk-go version 1.0.2

* publish jdcloud-sdk-go version 1.0.2

+++ services/baseanti/models/IpResourceFlowDetail.go
@@ -20,7 +20,7 @@ package models
 type IpResourceFlowDetail struct {
 
     /*  (Optional) */
-    Used []int `json:"used"`
+    Used []int64 `json:"used"`
 
     /*  (Optional) */
     Times []string `json:"times"`
